requirements.txt
beautifulsoup==4.3.2
readability-lxml==0.3.0.5
numpy==1.5.1
stemming==1.0
boto==2.34.0